Add equality checking beyond just instance identity for daisy objects
To match ruby's expectations, the eql? function checks runtime class
equality as well, but that causes an infinite recursion since the
runtime class inherits from the daisy object... so, for every built-in
ruby class that we use for interpreting the language, we need to also
override the equality functions appropriately.